# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Gym Project API
  version: 1.0.0
  description: JSON REST API interacting with the gym project backend
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-production-url.com/api
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: auth_key
      description: Authentication key obtained from /auth/login
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
...
    Post:
      type: object
      properties:
        postId:
          type: integer
        userId:
          type: integer
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    CreatePostRequest:
      type: object
      required:
        - title
        - content
      properties:
        user_id:
          type: integer
          example: 1
        title:
          type: string
          example: "Hello World"
        content:
          type: string
          example: "This is the post content"
    Session:
      type: object
      properties:
        session_id:
          type: integer
        trainer_id:
          type: integer
        session_date:
          type: string
          format: date
        session_time:
          type: string
        capacity:
          type: integer
...
paths:
...
  /posts:
    get:
      summary: Get all posts
      tags:
        - Posts
      security: []
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '500':
          description: Internal server error

    post:
      summary: Create a new post
      tags:
        - Posts
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '200':
          description: Post created successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /posts/me:
    get:
      summary: Get a post by user Id
      tags:
        - Posts
      security: 
        - ApiKeyAuth: []
      responses:
        '200':
          description: Post details
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Post not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /posts/{id}:
    delete:
      summary: Delete a post by ID
      tags:
        - Posts
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Post deleted successfully
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Post not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ==================== USERS ROUTES ====================
...
  /sessions:
    get:
      summary: Get all sessions
      tags:
        - Sessions
      security: []
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
        - in: query
          name: trainerId
          schema:
            type: integer
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description : Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

    post:
      summary: Create a new session
      tags:
        - Sessions
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Session'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Access denied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
          
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /sessions/{id}:
    get:
      summary: Get session by ID
      tags:
        - Sessions
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Session details
        '404':
          description: Session not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
          
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
          

    put:
      summary: Update session
      tags:
        - Sessions
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Session'
      responses:
        '200':
          description: Session updated
        '401':
          description: Unauthorized
        '403':
          description: Access denied
        '500':
          description: Internal server error

    delete:
      summary: Delete session
      tags:
        - Sessions
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Session deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Admins only
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
          

  # ==================== BOOKINGS ROUTES ====================
...
  /xml/sessions:
    get:
      summary: Export all sessions as XML
      tags:
        - Export
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Sessions exported as XML
          content:
            application/xml:
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Access denied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - ApiKeyAuth: []